requirements.txt
psycopg2-binary
requests
beautifulsoup4
ixbrlparse
openpyxl
lxml
